analyzer/pyproject.toml
[tool.poetry]
name = "analyzer"
version = "0.1.0"
description = ""
authors = ["nakatuba <nakatuba0626@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
fastapi = "^0.85.0"
fugashi = "^1.2.0"
torch = "^1.12.1"
transformers = "^4.22.1"
unidic-lite = "^1.0.8"
uvicorn = "^0.18.3"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
